python/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
rembg[cpu]>=2.0.72
Pillow>=10.4.0
numpy>=2.1.0
